Update the rockets page
