+++ src/assets/scss/rockets.module.scss
@@ -1,0 +1,12 @@
+.rockets {
+  width: 100%;
+
+  &__loading {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 1rem;
+  }
+}
