+++ src/components/Reserved.jsx
@@ -9,7 +9,7 @@ const Reserved = ({
   nameKey,
 }) => (
   <div className="flex-fill">
-    <h2>{title}</h2>
+    <h2 className="h2">{`My ${title}`}</h2>
     <div className={Styles.roundedborder}>
       {arr.length > 0 ? (
         <Table className="mb-0">
@@ -22,7 +22,7 @@ const Reserved = ({
           </tbody>
         </Table>
       ) : (
-        <p className="text-center">{`No reserved ${title}`}</p>
+        <p className="text-center pt-3 pb-4 ps-3 pe-3 mb-0">{`No reserved ${title}`}</p>
       )}
     </div>
   </div>
