+++ src/features/rockets/rocketsSlice.js
@@ -18,7 +18,7 @@ const rocketsSlice = createSlice({
   },
   reducers: {
     reserveRocketAction: (state, action) => {
-      const id = action.payload.id.toString();
+      const id = action.payload;
       const updatedRockets = state.rockets.map((rocket) => {
         if (rocket.id === id) {
           return {
